Feature/consumables (#170)
* Add: Quick Slot

#168

+ 퀵슬롯 UI 및 기능 구성
+ 소모성 아이템 사용 로직 구현
+ 소모성 아이템 사용을 위한 추가 변수 고안 <- 식량, 의약품
+ 의약품 사용 효과 적용

* Add: Consume Food

#168

+ 식량 아이템 사용

* Add: Replacing Item in Quick Slot

#168

~ 아이템 옮길 때 아래 아이템 있는지 체크할 때 y축 값 버그 수정
+ 퀵슬롯 아이템 바꾸기 구현

* Add: Quick Enroll for Quick Slot

#168 #169

+ 호버링 상태에서 LeftShift + 1, 2, 3 으로 퀵슬롯 1, 2, 3에 아이템 등록 기능 구현

+++ To-the-North-3d/Assets/Scripts/Users/Constants/ConditionConstraint.cs
@@ -6,7 +6,7 @@ namespace Assets.Scripts.Users
     /// 각 상태이상 별 효과 적용을 위한 델리게이트
     /// </summary>
     /// <param name="amount">값</param>
-    public delegate void ConditionControl(int amount);
+    public delegate void ConditionControl(float amount);
 
     public static class ConditionConstraint
     {
